is$ 처리 및 code넣어주기 및 데이터 넣기 작업

+++ src/index.js
@@ -13,10 +13,18 @@ const store = createStore(rootReducer);
 const path = `${PublicPath}/ppt/`; //경로 맞추어줌
 
 if (fs.existsSync(path)) {
+  let is$ = false;
   fs.readdirSync(path).forEach(file => {
-    let curPath = path + file;
-    fs.unlinkSync(curPath);
+    if (file.indexOf("$") !== -1) {
+      is$ = true;
+    }
   });
+  if (!is$) {
+    fs.readdirSync(path).forEach(file => {
+      let curPath = path + file;
+      fs.unlinkSync(curPath);
+    });
+  }
 }
 
 // index.html 에 작성한 ID 값
